Parse json client side

+++ Raspberry-Pi/src/webserver/http_response.py
@@ -5,6 +5,6 @@ from dataclasses import dataclass
 class HttpResponse:
     """ Represents a webserver response. """
     emergency: bool
-    angle: int
+    angle: float
     mode: str
     message: str
